+++ lib/src/analyzers/lint_analyzer/rules/rules_list/common_config.dart
@@ -1,0 +1,6 @@
+const functionalWidgetAnnotations = [
+  'FunctionalWidget',
+  'swidget',
+  'hwidget',
+  'hcwidget',
+];